+++ src/main/java/ca/corefacility/bioinformatics/irida/repositories/remote/resttemplate/SequenceFileMessageConverter.java
@@ -20,16 +20,12 @@ import org.springframework.http.converter.HttpMessageNotWritableException;
 import com.google.common.collect.Lists;
 
 /**
- * Message converter for converting application/fastq HTTP responses to a Java
+ * Message markdownConverter for converting application/fastq HTTP responses to a Java
  * Path temporary file
- * 
- *
  */
 public class SequenceFileMessageConverter implements HttpMessageConverter<Path> {
-	private static final Logger logger = LoggerFactory.getLogger(SequenceFileMessageConverter.class);
-
 	public static final MediaType MEDIA_TYPE = new MediaType("application","fastq");
-
+	private static final Logger logger = LoggerFactory.getLogger(SequenceFileMessageConverter.class);
 	private final String fileName;
 
 	public SequenceFileMessageConverter(String fileName) {
@@ -41,9 +37,8 @@ public class SequenceFileMessageConverter implements HttpMessageConverter<Path> 
 	 */
 	@Override
 	public boolean canRead(Class<?> clazz, MediaType mediaType) {
-		logger.trace("Testing converter for class " + clazz.getName() + " and mediatype " + mediaType);
-		if (mediaType != null && MEDIA_TYPE.includes(mediaType)
-				&& clazz.equals(Path.class)) {
+		logger.trace("Testing markdownConverter for class " + clazz.getName() + " and mediatype " + mediaType);
+		if (mediaType != null && MEDIA_TYPE.includes(mediaType) && clazz.equals(Path.class)) {
 			logger.trace("SequenceFileMessageConverter can read this message");
 			return true;
 		}