+++ src/main/java/ca/corefacility/bioinformatics/irida/ria/web/models/tables/TableRequest.java
@@ -1,4 +1,4 @@
-package ca.corefacility.bioinformatics.irida.ria.web.models;
+package ca.corefacility.bioinformatics.irida.ria.web.models.tables;
 
 import org.springframework.data.domain.Sort;
 
@@ -11,6 +11,9 @@ public class TableRequest {
 	private String sortColumn;
 	private String sortDirection;
 	private String search;
+
+	public TableRequest() {
+	}
 
 	public int getCurrent() {
 		return current;
@@ -40,15 +43,22 @@ public class TableRequest {
 		return search;
 	}
 
+	/**
+	 * Set the search term for the TableRequest.  This method will trim any leading and trailing whitespace.
+	 *
+	 * @param search The search term for the request
+	 */
 	public void setSearch(String search) {
-		this.search = search;
+		if (search != null) {
+			this.search = search.trim();
+		} else {
+			search = null;
+		}
 	}
 
 	/**
-	 * Since we he need an actual {@link Sort} object and cannot pass this from
-	 * the client, we create one from the information fathered from the client
-	 * Direction of sort
-	 * Column (attribute) of sort
+	 * Since we he need an actual {@link Sort} object and cannot pass this from the client, we create one from the
+	 * information fathered from the client Direction of sort Column (attribute) of sort
 	 *
 	 * @return {@link Sort}
 	 */