+++ src/main/webapp/resources/js/apis/analysis/analysis.js
@@ -83,3 +83,11 @@ export async function fetchAllPipelinesTypes() {
 export async function deleteAnalysisSubmissions({ ids }) {
   return axios.delete(`${ANALYSES_URL}/delete?ids=${ids.join(",")}`);
 }
+
+/**
+ * Fetch the current state of the analysis server.
+ * @return {Promise<T>} return a map of the running an queued counts.
+ */
+export async function fetchAnalysesQueueCounts() {
+  return axios.get("ajax/analyses/queue").then(({ data }) => data);
+}